requirements.txt
pytest
requests
psycopg2-binary
pytest-html
